+++ src/Env/Builder/EnvBuilderInterface.php
@@ -1,16 +1,17 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace LDL\Env\Builder;
 
-use LDL\Env\File\Finder\Exception\NoFilesFoundException;
+use LDL\Env\Util\File\Exception\ReadEnvFileException;
 use LDL\Env\Util\Line\Collection\EnvLineCollectionInterface;
+use LDL\File\Collection\ReadableFileCollection;
 
 interface EnvBuilderInterface
 {
     /**
-     * @return EnvLineCollectionInterface
-     * @throws NoFilesFoundException
+     * @throws ReadEnvFileException
      */
-    public  function build(): EnvLineCollectionInterface;
-
+    public function build(ReadableFileCollection $files): EnvLineCollectionInterface;
 }