+++ src/Env/Console/Command/PrintFilesCommand.php
@@ -1,11 +1,13 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace LDL\Env\Console\Command;
 
 use LDL\Env\File\Finder\EnvFileFinder;
-use Symfony\Component\Console\Command\Command as SymfonyCommand;
-use LDL\Env\File\Finder\Exception\NoFilesFoundException;
 use LDL\Env\File\Finder\Options\EnvFileFinderOptions;
+use Symfony\Component\Console\Command\Command as SymfonyCommand;
+use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -14,7 +16,7 @@ class PrintFilesCommand extends SymfonyCommand
 {
     public const COMMAND_NAME = 'env:print';
 
-    public function configure() : void
+    public function configure(): void
     {
         $defaults = EnvFileFinderOptions::fromArray([]);
 
@@ -23,15 +25,13 @@ class PrintFilesCommand extends SymfonyCommand
 
         $this->setName(self::COMMAND_NAME)
             ->setDescription('Prints .env files')
-            ->addOption(
+            ->addArgument(
                 'scan-directories',
-                'd',
-                InputOption::VALUE_REQUIRED,
+                InputArgument::REQUIRED,
                 sprintf(
                     'Comma separated list of directories to scan, default: %s',
                     $defaultDirectories
-                ),
-                $defaultDirectories
+                )
             )
             ->addOption(
                 'scan-files',
@@ -48,45 +48,31 @@ class PrintFilesCommand extends SymfonyCommand
     public function execute(InputInterface $input, OutputInterface $output)
     {
         try {
-            $this->printFiles($input, $output);
-            return 0;
-        }catch(\Exception $e){
-            $output->writeln("<error>{$e->getMessage()}</error>");
-            return 1;
-        }
-    }
+            $total = 0;
 
-    private function printFiles(
-        InputInterface $input,
-        OutputInterface $output
-    ) : void
-    {
-        $total = 0;
+            $output->writeln("<info>[ Env files list ]</info>\n");
 
-        $output->writeln("<info>[ Env files list ]</info>\n");
-
-        try{
             $finderOptions = EnvFileFinderOptions::fromArray([
-                'directories' => explode(',', $input->getOption('scan-directories')),
-                'files' => explode(',', $input->getOption('scan-files'))
+                'directories' => explode(',', $input->getArgument('scan-directories')),
+                'files' => explode(',', $input->getOption('scan-files')),
             ]);
 
             $finder = new EnvFileFinder($finderOptions);
 
             $files = $finder->find();
 
-        }catch(NoFilesFoundException $e){
-            $output->writeln("\n<error>{$e->getMessage()}</error>\n");
+            foreach ($files as $file) {
+                $total++;
+                $output->writeln($file);
+            }
 
-            return;
+            $output->writeln("\n<info>Total files found: $total</info>");
+
+            return self::SUCCESS;
+        } catch (\Exception $e) {
+            $output->writeln("<error>{$e->getMessage()}</error>");
+
+            return self::FAILURE;
         }
-
-        foreach($files as $file){
-            $total++;
-            $output->writeln($file);
-        }
-
-        $output->writeln("\n<info>Total files: $total</info>");
     }
-
 }