+++ src/Env/File/Finder/EnvFileFinderInterface.php
@@ -1,4 +1,6 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace LDL\Env\File\Finder;
 
@@ -6,15 +8,7 @@ use LDL\File\Collection\ReadableFileCollection;
 
 interface EnvFileFinderInterface
 {
-    /**
-     * @param bool $cache
-     * @return ReadableFileCollection
-     * @throws Exception\NoFilesFoundException
-     */
-    public function find(bool $cache = true) : ReadableFileCollection;
+    public function find(bool $cache = true): ReadableFileCollection;
 
-    /**
-     * @return Options\EnvFileFinderOptionsInterface
-     */
     public function getOptions(): Options\EnvFileFinderOptionsInterface;
 }